+++ ago-control-vaadin-site/src/main/java/org/agocontrol/client/EventProcessor.java
@@ -161,7 +161,7 @@ public class EventProcessor {
                         RecordSetDao.saveRecordSets(entityManager, Collections.singletonList(recordSet));
                     }
 
-                    final BigDecimal value = new BigDecimal(valueString);
+                    final Double value = new Double(valueString);
 
                     RecordDao.saveRecords(entityManager, Collections.singletonList(new Record(
                             owner,
